Added headerfiles and typecasts required for proper compilation with the pgi compiler.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@76 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/md_io/CheckpointWriter.cpp
@@ -37,7 +37,7 @@ void CheckpointWriter::doOutput(ParticleContainer* particleContainer,
     if(_incremental) {
       unsigned long temp = simstep/_writeFrequency;
       filenamestream << "-";
-      while(temp < floor(_numberOfTimesteps/_writeFrequency)){
+      while(temp < floor((double) (_numberOfTimesteps/_writeFrequency))){
   filenamestream << "0";
   temp = temp*10;
       }
