+++ src/md_io/VISWriter.cpp
@@ -36,7 +36,7 @@ void VISWriter::doOutput(ParticleContainer* particleContainer,
       if(_incremental) {
          unsigned long temp = simstep/_writeFrequency;
 	 filenamestream << "-";
-         while(temp < floor(_numberOfTimesteps/_writeFrequency)){
+         while(temp < floor((double) (_numberOfTimesteps/_writeFrequency))){
             filenamestream << "0";
             temp = temp*10;
          }
