+++ src/md_io/XyzWriter.cpp
@@ -38,7 +38,7 @@ void XyzWriter::doOutput(ParticleContainer* particleContainer,
     if(_incremental) {
       unsigned long temp = simstep/_writeFrequency;
       filenamestream << "-";
-      while(temp < floor(_numberOfTimesteps/_writeFrequency)){ 
+      while(temp < floor((double) (_numberOfTimesteps/_writeFrequency))){ 
         filenamestream << "0";
         temp = temp*10;
       }
